+++ src/JointPublisherROS.cpp
@@ -7,13 +7,13 @@ JointPublisherROS::JointPublisherROS(const std::string topic) : n("~") {
     pub = n.advertise<sensor_msgs::JointState>(topic, 1);
 }
 
-void JointPublisherROS::publish(const dart::Pose &pose, const float time) const {
+void JointPublisherROS::publish(const dart::Pose &pose, const double time) const {
     sensor_msgs::JointState js;
 
     js.header.stamp = ros::Time(time);
-    for(uint j(0); j<pose.getReducedArticulatedDimensions(); j++) {
+    for(int j(0); j<pose.getReducedArticulatedDimensions(); j++) {
         js.name.push_back(pose.getReducedName(j));
-        js.position.push_back(pose.getReducedArticulation()[j]);
+        js.position.push_back(double(pose.getReducedArticulation()[j]));
     }
 
     pub.publish(js);
